Emubase and Util minor fixes and cleanup.

+++ emulator/util/BitmapFile.cpp
@@ -102,11 +102,11 @@ bool BmpFile_SaveScreenshot(
         uint32_t rgb = *psrc;
         psrc++;
         uint8_t color = 0;
-        for (uint8_t c = 0; c < 256; c++)
+        for (int c = 0; c < 256; c++)
         {
             if (palette256[c] == rgb)
             {
-                color = c;
+                color = (uint8_t)c;
                 break;
             }
         }
@@ -282,11 +282,11 @@ bool PngFile_WriteImageData8(FILE * fpFile, uint32_t framenum, const uint32_t* p
         {
             uint32_t rgb = *(psrc++);
             uint8_t color = 0;
-            for (uint8_t c = 0; c < 256; c++)
+            for (int c = 0; c < 256; c++)
             {
                 if (palette256[c] == rgb)
                 {
-                    color = c;
+                    color = (uint8_t)c;
                     break;
                 }
             }
